feat: [Variables/Frontend] moved sort function into Tools, reduced brightness of normal text color
