+++ frontend/src/App.css
@@ -2,8 +2,9 @@
 .dark {
     --background-color: #202020;
     --textarea-background-color: #000;
-    --text-color: #ffffff;
+    --text-color: #d0d0d0;
     --text-color-less: #b0b0b0;
+    --text-color-headline: #FFFFFF;
     --text-color-label: #80b0FF;
     --border-color: #606060;
     --button-background-color: #404040;
@@ -25,6 +26,7 @@
     --textarea-background-color: #fff;
     --text-color: #000000;
     --text-color-less: #505050;
+    --text-color-headline: #300000;
     --text-color-label: #5080FF;
     --border-color: #a0a0a0;
     --button-background-color: #f0f0f0;
@@ -50,6 +52,7 @@ body {
 h1 {
     display: inline-block;
     padding: 5px;
+    color: var(--text-color-headline);
     text-shadow:
              5px  5px 9px var(--background-color, #FFFFFF),
             -5px  5px 9px var(--background-color, #FFFFFF),
